Retrieve lambda's secrets from Secrets Manager and store in file
The extension will examine its environment (which is shared with the
lambda function) and locate any variables which begin with "SECRET_."
It will then decrypt the lambda's SM secret, which is a collection
of key:value pairs which map to these environment variables (sans
the "SECRET_" prefix). For example, SECRET_PARAM_ONE will map to
PARAM_ONE. It will then store these k/v pairs in the /tmp/variables
file which is accessible to the lambda itself.